Added event `applyLink` for issue change default target for all links #841 Issue: https://github.com/xdan/jodit/issues/841
