+++ src/plugins/link/link.ts
@@ -8,15 +8,7 @@
  * @module plugins/link
  */
 
-import type {
-	IDictionary,
-	IJodit,
-	IControlType,
-	IUIForm,
-	IUIOption,
-	Nullable
-} from 'jodit/types';
-import { Config } from 'jodit/config';
+import type { IDictionary, IJodit, IUIForm, Nullable } from 'jodit/types';
 import { Dom } from 'jodit/core/dom';
 import {
 	attr,
@@ -26,114 +18,11 @@ import {
 	refs,
 	stripTags
 } from 'jodit/core/helpers';
-import { formTemplate } from './template';
 import { Plugin } from 'jodit/core/plugin';
 import { autobind } from 'jodit/core/decorators';
 import { Dialog, UIForm } from '../../modules';
 
-declare module 'jodit/config' {
-	interface Config {
-		link: {
-			/**
-			 * Template for the link dialog form
-			 */
-			formTemplate: (editor: IJodit) => string | HTMLElement | IUIForm;
-			formClassName?: string;
-
-			/**
-			 * Follow link address after dblclick
-			 */
-			followOnDblClick: boolean;
-
-			/**
-			 * Replace inserted youtube/vimeo link to `iframe`
-			 */
-			processVideoLink: boolean;
-
-			/**
-			 * Wrap inserted link
-			 */
-			processPastedLink: boolean;
-
-			/**
-			 * Show `no follow` checkbox in link dialog.
-			 */
-			noFollowCheckbox: boolean;
-
-			/**
-			 * Show `Open in new tab` checkbox in link dialog.
-			 */
-			openInNewTabCheckbox: boolean;
-
-			/**
-			 * Use an input text to ask the classname or a select or not ask
-			 */
-			modeClassName: 'input' | 'select';
-
-			/**
-			 * Allow multiple choises (to use with modeClassName="select")
-			 */
-			selectMultipleClassName: boolean;
-
-			/**
-			 * The size of the select (to use with modeClassName="select")
-			 */
-			selectSizeClassName?: number;
-
-			/**
-			 * The list of the option for the select (to use with modeClassName="select")
-			 */
-			selectOptionsClassName: IUIOption[];
-
-			hotkeys: string[];
-		};
-	}
-}
-
-Config.prototype.link = {
-	formTemplate,
-	followOnDblClick: false,
-	processVideoLink: true,
-	processPastedLink: true,
-	noFollowCheckbox: true,
-	openInNewTabCheckbox: true,
-	modeClassName: 'input',
-	selectMultipleClassName: true,
-	selectSizeClassName: 3,
-	selectOptionsClassName: [],
-	hotkeys: ['ctrl+k', 'cmd+k']
-};
-
-Config.prototype.controls.unlink = {
-	exec: (editor: IJodit, current: Node) => {
-		const anchor: HTMLAnchorElement | false = Dom.closest(
-			current,
-			'a',
-			editor.editor
-		) as HTMLAnchorElement;
-
-		if (anchor) {
-			Dom.unwrap(anchor);
-		}
-
-		editor.synchronizeValues();
-		editor.e.fire('hidePopup');
-	},
-	tooltip: 'Unlink'
-} as IControlType;
-
-Config.prototype.controls.link = {
-	isActive: (editor: IJodit): boolean => {
-		const current = editor.s.current();
-		return Boolean(current && Dom.closest(current, 'a', editor.editor));
-	},
-
-	popup: (editor: IJodit, current, self: IControlType, close: () => void) => {
-		return editor.e.fire('generateLinkForm.link', current, close);
-	},
-	tags: ['a'],
-	tooltip: 'Insert link'
-} as IControlType;
+import './config';
 
 /**
  * Process link. Insert, dblclick or remove format
@@ -158,6 +47,7 @@ export class link extends Plugin {
 		}
 
 		jodit.e.on('generateLinkForm.link', this.generateForm);
+
 		jodit.registerCommand('openLinkDialog', {
 			exec: () => {
 				const dialog = new Dialog({
@@ -485,6 +375,8 @@ export class link extends Plugin {
 						nofollow_checkbox.checked ? 'nofollow' : null
 					);
 				}
+
+				jodit.e.fire('applyLink', jodit, a, form);
 			});
 
 			jodit.synchronizeValues();
